requirements.txt
praw
python-dotenv
jinja2
weasyprint
yake
